Read all test configs from xctestrun file
